finished e2e test

+++ rating-service/src/main/java/by/shestakov/ratingservice/dto/response/AverageRatingResponse.java
@@ -1,8 +1,6 @@
 package by.shestakov.ratingservice.dto.response;
 
-import java.math.BigDecimal;
-
 public record AverageRatingResponse(
-        BigDecimal average
+        Double average
 ) {
 }
